refactor: moves and renames files to match new organization scheme
