+++ src/store/reducers/index.js
@@ -1,6 +1,6 @@
 import {combineReducers} from 'redux'
 import user from './user'
-import contactsArr from './contacts'
+import contactsArr from './contactsArr'
 import contactsHash from './contactsHash'
 
 export default combineReducers({user, contactsArr, contactsHash})
